backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
agno
google-generativeai
googlesearch-python
pycountry
asyncpg
psycopg2-binary
sqlalchemy[asyncio]
python-dotenv
